fix: FindDevContainer tries to lookup the pod and returns first result
So far we were waiting for the pod to come up in the implementation of
the `FindDevContainer` interface. This leads to a state where users
would have to wait for up to `options.PodTimeout` to delete a workspace
when the pod is stuck in an unschedulable/otherwise pending state.

`FindDevContainer` now behaves similar to implementations in the
dockerDriver in that it just looks up the pod and reports a result
immediately